final - Epic 6 and 7 implemented and few changes done in Entities

+++ BackEnd/RewardRedemptionApp/src/main/java/com/tcs/RewardRedemptionApp/Entity/Transactions.java
@@ -15,7 +15,7 @@ public class Transactions {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer transactionID;
+    private Long transactionID;
 
     private LocalDate transactionDate;
     private String transactionDetails;
